consegna/requirements.txt
torchvision~=0.17.2
matplotlib~=3.8.3
tqdm~=4.66.4
torchsummary~=1.5.1
scikit-learn~=1.4.1.post1
opencv-python~=4.9.0.80
numpy~=1.26.4
pandas~=2.2.1